Separating complexgroup tests into separate files
