+++ test/autorest/complexgroup/common.go
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+package complexgrouptest
+
+import (
+	"reflect"
+	"testing"
+)
+
+func deepEqualOrFatal(t *testing.T, result interface{}, expected interface{}) {
+	if !reflect.DeepEqual(result, expected) {
+		t.Fatalf("got %+v, want %+v", result, expected)
+	}
+}
